Refactoring common native code analysis
