+++ trueseeing/sig/android/crypto.py
@@ -8,7 +8,7 @@ from functools import cache
 
 from trueseeing.core.android.model import InvocationPattern
 from trueseeing.core.android.model import SignatureMixin
-from trueseeing.core.android.analysis.flow import DataFlow
+from trueseeing.core.android.analyze.flow import DataFlow
 
 if TYPE_CHECKING:
   from typing import Dict, Iterable, Optional, Any
