+++ trueseeing/sig/android/privacy.py
@@ -4,7 +4,7 @@ from typing import TYPE_CHECKING
 import re
 
 from trueseeing.core.android.model import InvocationPattern
-from trueseeing.core.android.analysis.flow import DataFlow
+from trueseeing.core.android.analyze.flow import DataFlow
 from trueseeing.core.android.model import SignatureMixin
 
 if TYPE_CHECKING:
